Network.AWS.AWSHealth.DescribeAffectedEntities

Description

Returns a list of entities that have been affected by the specified events, based on the specified filter criteria. Entities can refer to individual customer resources, groups of customer resources, or any other construct, depending on the AWS service. Events that have impact beyond that of the affected entities, or where the extent of impact is unknown, include at least one entity indicating this.

At least one event ARN is required. Results are sorted by the lastUpdatedTime of the entity, starting with the most recent.

This operation returns paginated results.

Creating a Request

describeAffectedEntities

Arguments

:: EntityFilter
daeFilter

-> DescribeAffectedEntities

Creates a value of DescribeAffectedEntities with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daeLocale - The locale (language) to return information in. English (en) is the default and the only supported value at this time.

daeNextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

daeMaxResults - The maximum number of items to return in one batch, between 10 and 100, inclusive.

daeFilter - Values to narrow the results returned. At least one event ARN is required.

data DescribeAffectedEntities

See: describeAffectedEntities smart constructor.

Request Lenses

daeLocale :: Lens' DescribeAffectedEntities (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

daeNextToken :: Lens' DescribeAffectedEntities (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

daeMaxResults :: Lens' DescribeAffectedEntities (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

daeFilter :: Lens' DescribeAffectedEntities EntityFilter

Values to narrow the results returned. At least one event ARN is required.

Destructuring the Response

describeAffectedEntitiesResponse

Arguments

:: Int
daersResponseStatus

-> DescribeAffectedEntitiesResponse

Creates a value of DescribeAffectedEntitiesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daersEntities - The entities that match the filter criteria.

daersNextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

daersResponseStatus - -- | The response status code.

data DescribeAffectedEntitiesResponse

See: describeAffectedEntitiesResponse smart constructor.

Response Lenses

daersEntities :: Lens' DescribeAffectedEntitiesResponse [AffectedEntity]

The entities that match the filter criteria.

daersNextToken :: Lens' DescribeAffectedEntitiesResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

daersResponseStatus :: Lens' DescribeAffectedEntitiesResponse Int

- | The response status code.