Network.AWS.AWSHealth.DescribeEventDetails

Description

Returns detailed information about one or more specified events. Information includes standard event data (region, service, etc., as returned by DescribeEvents ), a detailed event description, and possible additional metadata that depends upon the nature of the event. Affected entities are not included; to retrieve those, use the DescribeAffectedEntities operation.

If a specified event cannot be retrieved, an error message is returned for that event.

Creating a Request

describeEventDetails

Arguments

:: NonEmpty Text
dedEventARNs

-> DescribeEventDetails

Creates a value of DescribeEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dedLocale - The locale (language) to return information in. English (en) is the default and the only supported value at this time.

dedEventARNs - A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::eventAWS_EC2_MAINTENANCE_5331", "arn:aws:health:us-west-1::eventAWS_EBS_LOST_VOLUME_xyz"

data DescribeEventDetails

See: describeEventDetails smart constructor.

Request Lenses

dedLocale :: Lens' DescribeEventDetails (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

dedEventARNs :: Lens' DescribeEventDetails (NonEmpty Text)

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::eventAWS_EC2_MAINTENANCE_5331", "arn:aws:health:us-west-1::eventAWS_EBS_LOST_VOLUME_xyz"

Destructuring the Response

describeEventDetailsResponse

Arguments

:: Int
dedrsResponseStatus

-> DescribeEventDetailsResponse

Creates a value of DescribeEventDetailsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dedrsSuccessfulSet - Information about the events that could be retrieved.

dedrsFailedSet - Error messages for any events that could not be retrieved.

dedrsResponseStatus - -- | The response status code.

data DescribeEventDetailsResponse

See: describeEventDetailsResponse smart constructor.

Response Lenses

dedrsSuccessfulSet :: Lens' DescribeEventDetailsResponse [EventDetails]

Information about the events that could be retrieved.

dedrsFailedSet :: Lens' DescribeEventDetailsResponse [EventDetailsErrorItem]

Error messages for any events that could not be retrieved.

dedrsResponseStatus :: Lens' DescribeEventDetailsResponse Int

- | The response status code.