Network.AWS.AWSHealth.DescribeEventTypes

Description

Returns the event types that meet the specified filter criteria. If no filter criteria are specified, all event types are returned, in no particular order.

This operation returns paginated results.

Creating a Request

describeEventTypes :: DescribeEventTypes

Creates a value of DescribeEventTypes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

detLocale - The locale (language) to return information in. English (en) is the default and the only supported value at this time.

detNextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

detFilter - Values to narrow the results returned.

detMaxResults - The maximum number of items to return in one batch, between 10 and 100, inclusive.

data DescribeEventTypes

See: describeEventTypes smart constructor.

Request Lenses

detLocale :: Lens' DescribeEventTypes (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

detNextToken :: Lens' DescribeEventTypes (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

detFilter :: Lens' DescribeEventTypes (Maybe EventTypeFilter)

Values to narrow the results returned.

detMaxResults :: Lens' DescribeEventTypes (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

Destructuring the Response

describeEventTypesResponse

Arguments

:: Int
detrsResponseStatus

-> DescribeEventTypesResponse

Creates a value of DescribeEventTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

detrsEventTypes - A list of event types that match the filter criteria. Event types have a category (issue , accountNotification , or scheduledChange ), a service (for example, EC2 , RDS , DATAPIPELINE , BILLING ), and a code (in the format AWS_SERVICE _DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT ).

detrsNextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

detrsResponseStatus - -- | The response status code.

data DescribeEventTypesResponse

See: describeEventTypesResponse smart constructor.

Response Lenses

detrsEventTypes :: Lens' DescribeEventTypesResponse [EventType]

A list of event types that match the filter criteria. Event types have a category (issue , accountNotification , or scheduledChange ), a service (for example, EC2 , RDS , DATAPIPELINE , BILLING ), and a code (in the format AWS_SERVICE _DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT ).

detrsNextToken :: Lens' DescribeEventTypesResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

detrsResponseStatus :: Lens' DescribeEventTypesResponse Int

- | The response status code.