Network.AWS.AWSHealth.DescribeEvents

Description

Returns information about events that meet the specified filter criteria. Events are returned in a summary form and do not include the detailed description, any additional metadata that depends on the event type, or any affected resources. To retrieve that information, use the DescribeEventDetails and DescribeAffectedEntities operations.

If no filter criteria are specified, all events are returned. Results are sorted by lastModifiedTime , starting with the most recent.

This operation returns paginated results.

Creating a Request

describeEvents :: DescribeEvents

Creates a value of DescribeEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deLocale - The locale (language) to return information in. English (en) is the default and the only supported value at this time.

deNextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

deFilter - Values to narrow the results returned.

deMaxResults - The maximum number of items to return in one batch, between 10 and 100, inclusive.

data DescribeEvents

See: describeEvents smart constructor.

Request Lenses

deLocale :: Lens' DescribeEvents (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

deNextToken :: Lens' DescribeEvents (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

deFilter :: Lens' DescribeEvents (Maybe EventFilter)

Values to narrow the results returned.

deMaxResults :: Lens' DescribeEvents (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

Destructuring the Response

describeEventsResponse

Arguments

:: Int
dersResponseStatus

-> DescribeEventsResponse

Creates a value of DescribeEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dersNextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

dersEvents - The events that match the specified filter criteria.

dersResponseStatus - -- | The response status code.

data DescribeEventsResponse

See: describeEventsResponse smart constructor.

Response Lenses

dersNextToken :: Lens' DescribeEventsResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

dersEvents :: Lens' DescribeEventsResponse [Event]

The events that match the specified filter criteria.

dersResponseStatus :: Lens' DescribeEventsResponse Int

- | The response status code.