Network.AWS.AWSHealth.Types

Description

Service Configuration

awsHealth :: Service

API version 2016-08-04 of the Amazon Health APIs and Notifications SDK configuration.

Errors

_InvalidPaginationToken :: AsError a => Getting (First ServiceError) a ServiceError

The specified pagination token (nextToken ) is not valid.

_UnsupportedLocale :: AsError a => Getting (First ServiceError) a ServiceError

The specified locale is not supported.

EntityStatusCode

data EntityStatusCode

Constructors

Impaired

Unimpaired

Unknown

EventAggregateField

data EventAggregateField

Constructors

EventTypeCategory

EventStatusCode

data EventStatusCode

Constructors

Closed

Open

Upcoming

EventTypeCategory

data EventTypeCategory

Constructors

AccountNotification

Issue

ScheduledChange

AffectedEntity

data AffectedEntity

Information about an entity that is affected by a Health event.

See: affectedEntity smart constructor.

affectedEntity :: AffectedEntity

Creates a value of AffectedEntity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aeLastUpdatedTime - The most recent time that the entity was updated.

aeEntityValue - The ID of the affected entity.

aeAwsAccountId - The 12-digit AWS account number that contains the affected entity.

aeEventARN - The unique identifier for the event. Format: arn:aws:health:event-region ::event/EVENT_TYPE_PLUS_ID . Example: arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331

aeEntityARN - The unique identifier for the entity. Format: arn:aws:health:entity-region :aws-account :entity/entity-id . Example: arn:aws:health:us-east-1:111222333444:entity/AVh5GGT7ul1arKr1sE1K

aeTags - A map of entity tags attached to the affected entity.

aeStatusCode - The most recent status of the entity affected by the event. The possible values are IMPAIRED , UNIMPAIRED , and UNKNOWN .

aeLastUpdatedTime :: Lens' AffectedEntity (Maybe UTCTime)

The most recent time that the entity was updated.

aeEntityValue :: Lens' AffectedEntity (Maybe Text)

The ID of the affected entity.

aeAwsAccountId :: Lens' AffectedEntity (Maybe Text)

The 12-digit AWS account number that contains the affected entity.

aeEventARN :: Lens' AffectedEntity (Maybe Text)

The unique identifier for the event. Format: arn:aws:health:event-region ::event/EVENT_TYPE_PLUS_ID . Example: arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331

aeEntityARN :: Lens' AffectedEntity (Maybe Text)

The unique identifier for the entity. Format: arn:aws:health:entity-region :aws-account :entity/entity-id . Example: arn:aws:health:us-east-1:111222333444:entity/AVh5GGT7ul1arKr1sE1K

aeTags :: Lens' AffectedEntity (HashMap Text Text)

A map of entity tags attached to the affected entity.

aeStatusCode :: Lens' AffectedEntity (Maybe EntityStatusCode)

The most recent status of the entity affected by the event. The possible values are IMPAIRED , UNIMPAIRED , and UNKNOWN .

DateTimeRange

data DateTimeRange

A range of dates and times that is used by the EventFilter and EntityFilter objects. If from is set and to is set: match items where the timestamp (startTime , endTime , or lastUpdatedTime ) is between from and to inclusive. If from is set and to is not set: match items where the timestamp value is equal to or after from . If from is not set and to is set: match items where the timestamp value is equal to or before to .

See: dateTimeRange smart constructor.

dateTimeRange :: DateTimeRange

Creates a value of DateTimeRange with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrTo - The ending date and time of a time range.

dtrFrom - The starting date and time of a time range.

dtrTo :: Lens' DateTimeRange (Maybe UTCTime)

The ending date and time of a time range.

dtrFrom :: Lens' DateTimeRange (Maybe UTCTime)

The starting date and time of a time range.

EntityAggregate

data EntityAggregate

The number of entities that are affected by one or more events. Returned by the DescribeEntityAggregates operation.

See: entityAggregate smart constructor.

entityAggregate :: EntityAggregate

Creates a value of EntityAggregate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eCount - The number entities that match the criteria for the specified events.

eEventARN - The unique identifier for the event. Format: arn:aws:health:event-region ::event/EVENT_TYPE_PLUS_ID . Example: arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331

eCount :: Lens' EntityAggregate (Maybe Int)

The number entities that match the criteria for the specified events.

eEventARN :: Lens' EntityAggregate (Maybe Text)

The unique identifier for the event. Format: arn:aws:health:event-region ::event/EVENT_TYPE_PLUS_ID . Example: arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331

EntityFilter

data EntityFilter

The values to use to filter results from the DescribeAffectedEntities operation.

See: entityFilter smart constructor.

entityFilter

Arguments

:: NonEmpty Text
eEventARNs

-> EntityFilter

Creates a value of EntityFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eStatusCodes - A list of entity status codes (IMPAIRED , UNIMPAIRED , or UNKNOWN ).

eEntityARNs - A list of entity ARNs (unique identifiers).

eEntityValues - A list of IDs for affected entities.

eTags - A map of entity tags attached to the affected entity.

eLastUpdatedTimes - A list of the most recent dates and times that the entity was updated.

eEventARNs - A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::eventAWS_EC2_MAINTENANCE_5331", "arn:aws:health:us-west-1::eventAWS_EBS_LOST_VOLUME_xyz"

eStatusCodes :: Lens' EntityFilter (Maybe (NonEmpty EntityStatusCode))

A list of entity status codes (IMPAIRED , UNIMPAIRED , or UNKNOWN ).

eEntityARNs :: Lens' EntityFilter (Maybe (NonEmpty Text))

A list of entity ARNs (unique identifiers).

eEntityValues :: Lens' EntityFilter (Maybe (NonEmpty Text))

A list of IDs for affected entities.

eTags :: Lens' EntityFilter [HashMap Text Text]

A map of entity tags attached to the affected entity.

eLastUpdatedTimes :: Lens' EntityFilter (Maybe (NonEmpty DateTimeRange))

A list of the most recent dates and times that the entity was updated.

eEventARNs :: Lens' EntityFilter (NonEmpty Text)

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::eventAWS_EC2_MAINTENANCE_5331", "arn:aws:health:us-west-1::eventAWS_EBS_LOST_VOLUME_xyz"

Event

data Event

Summary information about an event, returned by the DescribeEvents operation. The DescribeEventDetails operation also returns this information, as well as the EventDescription and additional event metadata.

See: event smart constructor.

event :: Event

Creates a value of Event with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eLastUpdatedTime - The most recent date and time that the event was updated.

eArn - The unique identifier for the event. Format: arn:aws:health:event-region ::event/EVENT_TYPE_PLUS_ID . Example: arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331

eService - The AWS service that is affected by the event. For example, EC2 , RDS .

eStartTime - The date and time that the event began.

eEventTypeCode - The unique identifier for the event type. The format is AWS_SERVICE _DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT .

eEventTypeCategory - The

eAvailabilityZone - The AWS Availability Zone of the event. For example, us-east-1a.

eEndTime - The date and time that the event ended.

eRegion - The AWS region name of the event.

eStatusCode - The most recent status of the event. Possible values are open , closed , and upcoming .

eLastUpdatedTime :: Lens' Event (Maybe UTCTime)

The most recent date and time that the event was updated.

eArn :: Lens' Event (Maybe Text)

The unique identifier for the event. Format: arn:aws:health:event-region ::event/EVENT_TYPE_PLUS_ID . Example: arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331

eService :: Lens' Event (Maybe Text)

The AWS service that is affected by the event. For example, EC2 , RDS .

eStartTime :: Lens' Event (Maybe UTCTime)

The date and time that the event began.

eEventTypeCode :: Lens' Event (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE _DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT .

eEventTypeCategory :: Lens' Event (Maybe EventTypeCategory)

The

eAvailabilityZone :: Lens' Event (Maybe Text)

The AWS Availability Zone of the event. For example, us-east-1a.

eEndTime :: Lens' Event (Maybe UTCTime)

The date and time that the event ended.

eRegion :: Lens' Event (Maybe Text)

The AWS region name of the event.

eStatusCode :: Lens' Event (Maybe EventStatusCode)

The most recent status of the event. Possible values are open , closed , and upcoming .

EventAggregate

data EventAggregate

The number of events of each issue type. Returned by the DescribeEventAggregates operation.

See: eventAggregate smart constructor.

eventAggregate :: EventAggregate

Creates a value of EventAggregate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eaCount - The number of events of the associated issue type.

eaAggregateValue - The issue type for the associated count.

eaCount :: Lens' EventAggregate (Maybe Int)

The number of events of the associated issue type.

eaAggregateValue :: Lens' EventAggregate (Maybe Text)

The issue type for the associated count.

EventDescription

data EventDescription

The detailed description of the event. Included in the information returned by the DescribeEventDetails operation.

See: eventDescription smart constructor.

eventDescription :: EventDescription

Creates a value of EventDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

edLatestDescription - The most recent description of the event.

edLatestDescription :: Lens' EventDescription (Maybe Text)

The most recent description of the event.

EventDetails

data EventDetails

Detailed information about an event. A combination of an Event object, an EventDescription object, and additional metadata about the event. Returned by the DescribeEventDetails operation.

See: eventDetails smart constructor.

eventDetails :: EventDetails

Creates a value of EventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

edEvent - Summary information about the event.

edEventDescription - The most recent description of the event.

edEventMetadata - Additional metadata about the event.

edEvent :: Lens' EventDetails (Maybe Event)

Summary information about the event.

edEventDescription :: Lens' EventDetails (Maybe EventDescription)

The most recent description of the event.

edEventMetadata :: Lens' EventDetails (HashMap Text Text)

Additional metadata about the event.

EventDetailsErrorItem

data EventDetailsErrorItem

Error information returned when a DescribeEventDetails operation cannot find a specified event.

See: eventDetailsErrorItem smart constructor.

eventDetailsErrorItem :: EventDetailsErrorItem

Creates a value of EventDetailsErrorItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

edeiEventARN - The unique identifier for the event. Format: arn:aws:health:event-region ::event/EVENT_TYPE_PLUS_ID . Example: arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331

edeiErrorName - The name of the error.

edeiErrorMessage - A message that describes the error.

edeiEventARN :: Lens' EventDetailsErrorItem (Maybe Text)

The unique identifier for the event. Format: arn:aws:health:event-region ::event/EVENT_TYPE_PLUS_ID . Example: arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331

edeiErrorName :: Lens' EventDetailsErrorItem (Maybe Text)

The name of the error.

edeiErrorMessage :: Lens' EventDetailsErrorItem (Maybe Text)

A message that describes the error.

EventFilter

data EventFilter

The values to use to filter results from the DescribeEvents and DescribeEventAggregates operations.

See: eventFilter smart constructor.

eventFilter :: EventFilter

Creates a value of EventFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

efEventARNs - A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::eventAWS_EC2_MAINTENANCE_5331", "arn:aws:health:us-west-1::eventAWS_EBS_LOST_VOLUME_xyz"

efEventTypeCategories - A list of event type category codes (issue , scheduledChange , or accountNotification ).

efEventTypeCodes - A list of unique identifiers for event types. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT,AWS_RDS_MAINTENANCE_SCHEDULED

efRegions - A list of AWS regions.

efEventStatusCodes - A list of event status codes.

efEndTimes - A list of dates and times that the event ended.

efAvailabilityZones - A list of AWS availability zones.

efEntityARNs - A list of entity ARNs (unique identifiers).

efEntityValues - A list of entity identifiers, such as EC2 instance IDs (i-34ab692e ) or EBS volumes (vol-426ab23e ).

efStartTimes - A list of dates and times that the event began.

efServices - The AWS services associated with the event. For example, EC2 , RDS .

efTags - A map of entity tags attached to the affected entity.

efLastUpdatedTimes - A list of dates and times that the event was last updated.

efEventARNs :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::eventAWS_EC2_MAINTENANCE_5331", "arn:aws:health:us-west-1::eventAWS_EBS_LOST_VOLUME_xyz"

efEventTypeCategories :: Lens' EventFilter (Maybe (NonEmpty EventTypeCategory))

A list of event type category codes (issue , scheduledChange , or accountNotification ).

efEventTypeCodes :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of unique identifiers for event types. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT,AWS_RDS_MAINTENANCE_SCHEDULED

efRegions :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of AWS regions.

efEventStatusCodes :: Lens' EventFilter (Maybe (NonEmpty EventStatusCode))

A list of event status codes.

efEndTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event ended.

efAvailabilityZones :: Lens' EventFilter [Text]

A list of AWS availability zones.

efEntityARNs :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of entity ARNs (unique identifiers).

efEntityValues :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of entity identifiers, such as EC2 instance IDs (i-34ab692e ) or EBS volumes (vol-426ab23e ).

efStartTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event began.

efServices :: Lens' EventFilter (Maybe (NonEmpty Text))

The AWS services associated with the event. For example, EC2 , RDS .

efTags :: Lens' EventFilter [HashMap Text Text]

A map of entity tags attached to the affected entity.

efLastUpdatedTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event was last updated.

EventType

data EventType

Metadata about a type of event that is reported by AWS Health. Data consists of the category (for example, issue ), the service (for example, EC2 ), and the event type code (for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT ).

See: eventType smart constructor.

eventType :: EventType

Creates a value of EventType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

etService - The AWS service that is affected by the event. For example, EC2 , RDS .

etCategory - A list of event type category codes (issue , scheduledChange , or accountNotification ).

etCode - The unique identifier for the event type. The format is AWS_SERVICE _DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT .

etService :: Lens' EventType (Maybe Text)

The AWS service that is affected by the event. For example, EC2 , RDS .

etCategory :: Lens' EventType (Maybe EventTypeCategory)

A list of event type category codes (issue , scheduledChange , or accountNotification ).

etCode :: Lens' EventType (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE _DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT .

EventTypeFilter

data EventTypeFilter

The values to use to filter results from the DescribeEventTypes operation.

See: eventTypeFilter smart constructor.

eventTypeFilter :: EventTypeFilter

Creates a value of EventTypeFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

etfEventTypeCategories - A list of event type category codes (issue , scheduledChange , or accountNotification ).

etfEventTypeCodes - A list of event type codes.

etfServices - The AWS services associated with the event. For example, EC2 , RDS .

etfEventTypeCategories :: Lens' EventTypeFilter (Maybe (NonEmpty EventTypeCategory))

A list of event type category codes (issue , scheduledChange , or accountNotification ).

etfEventTypeCodes :: Lens' EventTypeFilter (Maybe (NonEmpty Text))

A list of event type codes.

etfServices :: Lens' EventTypeFilter (Maybe (NonEmpty Text))

The AWS services associated with the event. For example, EC2 , RDS .